+++ src/main/java/mod/gottsch/forge/dungeonblocks/core/block/WeatheringHeavyTrapDoorBlock.java
@@ -1,14 +1,31 @@
+/*
+ * This file is part of  Dungeon Blocks.
+ * Copyright (c) 2025 Mark Gottschling (gottsch)
+ *
+ * All rights reserved.
+ *
+ * Dungeon Blocks is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Dungeon Blocks is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Dungeon Blocks.  If not, see <http://www.gnu.org/licenses/lgpl>.
+ */
 package mod.gottsch.forge.dungeonblocks.core.block;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
-import net.minecraft.world.level.block.WeatheringCopper;
 import net.minecraft.world.level.block.WeatheringCopper.WeatherState;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.BlockBehaviour.Properties;
 
-public class WeatheringHeavyTrapDoorBlock extends HeavyTrapDoorBlock implements WeatheringCopper {
+public class WeatheringHeavyTrapDoorBlock extends HeavyTrapDoorBlock implements ModWeatheringCopper {
    private final WeatherState weatherState;
 
    public WeatheringHeavyTrapDoorBlock(WeatherState weatherState, Properties properties) {
@@ -21,7 +38,7 @@ public class WeatheringHeavyTrapDoorBlock extends HeavyTrapDoorBlock implements 
    }
 
    public boolean isRandomlyTicking(BlockState state) {
-      return WeatheringCopper.getNext(state.getBlock()).isPresent();
+      return ModWeatheringCopper.getNext(state.getBlock()).isPresent();
    }
 
    public WeatherState getAge() {