+++ src/main/java/mod/gottsch/forge/dungeonblocks/datagen/ModBlockStateProvider.java
@@ -4,6 +4,7 @@ import mod.gottsch.forge.dungeonblocks.core.block.*;
 import mod.gottsch.forge.dungeonblocks.DungeonBlocks;
 import mod.gottsch.forge.dungeonblocks.core.setup.Registration;
 import mod.gottsch.forge.dungeonblocks.core.state.properties.FacadeShape;
+import mod.gottsch.forge.gottschcore.block.FacingHalfBlock;
 import mod.gottsch.forge.gottschcore.block.IFacingBlock;
 import net.minecraft.core.Direction;
 import net.minecraft.data.PackOutput;
@@ -11,6 +12,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.EnumProperty;
+import net.minecraft.world.level.block.state.properties.Half;
 import net.minecraft.world.level.block.state.properties.Property;
 import net.minecraftforge.client.model.generators.*;
 import net.minecraftforge.common.data.ExistingFileHelper;
@@ -112,8 +114,8 @@ public class ModBlockStateProvider extends BlockStateProvider {
         simpleSingleTexture(ModBlocks.WAXED_EXPOSED_COPPER_GRATE, modLoc("block/template_cube_cutout"), modLoc("block/exposed_copper_grate"));
         simpleSingleTexture(ModBlocks.WAXED_WEATHERED_COPPER_GRATE, modLoc("block/template_cube_cutout"), modLoc("block/weathered_copper_grate"));
         simpleSingleTexture(ModBlocks.WAXED_OXIDIZED_COPPER_GRATE, modLoc("block/template_cube_cutout"), modLoc("block/oxidized_copper_grate"));
-        heavyGrateBlock(ModBlocks.DARK_IRON_GRATE, modLoc("block/dark_iron"));
 
+        heavyGrateBlock(ModBlocks.DARK_IRON_GRATE, modLoc("block/dark_iron"));
         heavyGrateBlock(ModBlocks.COPPER_HEAVY_GRATE, mcLoc("block/copper_block"));
         heavyGrateBlock(ModBlocks.EXPOSED_COPPER_HEAVY_GRATE, mcLoc("block/exposed_copper"));
         heavyGrateBlock(ModBlocks.WEATHERED_COPPER_HEAVY_GRATE, mcLoc("block/weathered_copper"));
@@ -123,10 +125,53 @@ public class ModBlockStateProvider extends BlockStateProvider {
         heavyGrateBlock(ModBlocks.WAXED_WEATHERED_COPPER_HEAVY_GRATE, mcLoc("block/weathered_copper"));
         heavyGrateBlock(ModBlocks.WAXED_OXIDIZED_COPPER_HEAVY_GRATE, mcLoc("block/oxidized_copper"));
 
+        valveWheelBlock(ModBlocks.COPPER_VALVE_WHEEL, mcLoc("block/copper_block"));
+        valveWheelBlock(ModBlocks.EXPOSED_COPPER_VALVE_WHEEL, mcLoc("block/exposed_copper"));
+        valveWheelBlock(ModBlocks.WEATHERED_COPPER_VALVE_WHEEL, mcLoc("block/weathered_copper"));
+        valveWheelBlock(ModBlocks.OXIDIZED_COPPER_VALVE_WHEEL, mcLoc("block/oxidized_copper"));
+        valveWheelBlock(ModBlocks.WAXED_COPPER_VALVE_WHEEL, mcLoc("block/copper_block"));
+        valveWheelBlock(ModBlocks.WAXED_EXPOSED_COPPER_VALVE_WHEEL, mcLoc("block/exposed_copper"));
+        valveWheelBlock(ModBlocks.WAXED_WEATHERED_COPPER_VALVE_WHEEL, mcLoc("block/weathered_copper"));
+        valveWheelBlock(ModBlocks.WAXED_OXIDIZED_COPPER_VALVE_WHEEL, mcLoc("block/oxidized_copper"));
+
+
+        plateBracketBlock(ModBlocks.IRON_PLATE_BRACKET, modLoc("block/iron_plate"));
+        plateBracketBlock(ModBlocks.DARK_IRON_PLATE_BRACKET, modLoc("block/dark_iron"));
+        plateBracketBlock(ModBlocks.COPPER_PLATE_BRACKET, mcLoc("block/copper_block"));
+        plateBracketBlock(ModBlocks.EXPOSED_COPPER_PLATE_BRACKET, mcLoc("block/exposed_copper"));
+        plateBracketBlock(ModBlocks.WEATHERED_COPPER_PLATE_BRACKET, mcLoc("block/weathered_copper"));
+        plateBracketBlock(ModBlocks.OXIDIZED_COPPER_PLATE_BRACKET, mcLoc("block/oxidized_copper"));
+        plateBracketBlock(ModBlocks.WAXED_COPPER_PLATE_BRACKET, mcLoc("block/copper_block"));
+        plateBracketBlock(ModBlocks.WAXED_EXPOSED_COPPER_PLATE_BRACKET, mcLoc("block/exposed_copper"));
+        plateBracketBlock(ModBlocks.WAXED_WEATHERED_COPPER_PLATE_BRACKET, mcLoc("block/weathered_copper"));
+        plateBracketBlock(ModBlocks.WAXED_OXIDIZED_COPPER_PLATE_BRACKET, mcLoc("block/oxidized_copper"));
+
+        anglePlateBracketBlock(ModBlocks.DARK_IRON_ANGLE_PLATE_BRACKET, modLoc("block/dark_iron"));
+        anglePlateBracketBlock(ModBlocks.COPPER_ANGLE_PLATE_BRACKET, mcLoc("block/copper_block"));
+        anglePlateBracketBlock(ModBlocks.EXPOSED_COPPER_ANGLE_PLATE_BRACKET, mcLoc("block/exposed_copper"));
+        anglePlateBracketBlock(ModBlocks.WEATHERED_COPPER_ANGLE_PLATE_BRACKET, mcLoc("block/weathered_copper"));
+        anglePlateBracketBlock(ModBlocks.OXIDIZED_COPPER_ANGLE_PLATE_BRACKET, mcLoc("block/oxidized_copper"));
+
+        anglePlateBracketBlock(ModBlocks.WAXED_COPPER_ANGLE_PLATE_BRACKET, mcLoc("block/copper_block"));
+        anglePlateBracketBlock(ModBlocks.WAXED_EXPOSED_COPPER_ANGLE_PLATE_BRACKET, mcLoc("block/exposed_copper"));
+        anglePlateBracketBlock(ModBlocks.WAXED_WEATHERED_COPPER_ANGLE_PLATE_BRACKET, mcLoc("block/weathered_copper"));
+        anglePlateBracketBlock(ModBlocks.WAXED_OXIDIZED_COPPER_ANGLE_PLATE_BRACKET, mcLoc("block/oxidized_copper"));
+
+        cornerPlateBracketBlock(ModBlocks.DARK_IRON_CORNER_PLATE_BRACKET, modLoc("block/dark_iron"));
+        cornerPlateBracketBlock(ModBlocks.COPPER_CORNER_PLATE_BRACKET, mcLoc("block/copper_block"));
+        cornerPlateBracketBlock(ModBlocks.EXPOSED_COPPER_CORNER_PLATE_BRACKET, mcLoc("block/exposed_copper"));
+        cornerPlateBracketBlock(ModBlocks.WEATHERED_COPPER_CORNER_PLATE_BRACKET, mcLoc("block/weathered_copper"));
+        cornerPlateBracketBlock(ModBlocks.OXIDIZED_COPPER_CORNER_PLATE_BRACKET, mcLoc("block/oxidized_copper"));
+
+        cornerPlateBracketBlock(ModBlocks.WAXED_COPPER_CORNER_PLATE_BRACKET, mcLoc("block/copper_block"));
+        cornerPlateBracketBlock(ModBlocks.WAXED_EXPOSED_COPPER_CORNER_PLATE_BRACKET, mcLoc("block/exposed_copper"));
+        cornerPlateBracketBlock(ModBlocks.WAXED_WEATHERED_COPPER_CORNER_PLATE_BRACKET, mcLoc("block/weathered_copper"));
+        cornerPlateBracketBlock(ModBlocks.WAXED_OXIDIZED_COPPER_CORNER_PLATE_BRACKET, mcLoc("block/oxidized_copper"));
+
         wallRingBlock(ModBlocks.WALL_RING);
         hayPatchBlock(ModBlocks.HAY_PATCH);
         hayPatchBlock(ModBlocks.DIRTY_HAY_PATCH, modLoc("block/dirty_hay"));
-        plateBracketBlock(ModBlocks.PLATE_BRACKET_BLOCK);
+//        basedBlock(ModBlocks.ANGLE_PLATE_BRACKET_BLOCK, "angle_plate_bracket_block");
 
         sewerBlock(ModBlocks.WEATHERED_COPPER_SEWER, modLoc("block/weathered_copper_pipe"), mcLoc("block/weathered_copper"));
         sewerBlock(ModBlocks.TERRACOTTA_SEWER, mcLoc("block/terracotta"), mcLoc("block/terracotta"));
@@ -151,14 +196,26 @@ public class ModBlockStateProvider extends BlockStateProvider {
         simpleBlock(ModBlocks.MOSSY_SQUARE_STONE_BRICK.get());
         rectangleLeftHorizontalBlock(ModBlocks.LEFT_LARGE_STONE_BRICK, ModBlocks.RIGHT_LARGE_STONE_BRICK);
         rectangleRightHorizontalBlock(ModBlocks.RIGHT_LARGE_STONE_BRICK, ModBlocks.LEFT_LARGE_STONE_BRICK);
+        rectangleLeftHorizontalBlock(ModBlocks.MOSSY_LEFT_LARGE_STONE_BRICK, ModBlocks.MOSSY_RIGHT_LARGE_STONE_BRICK);
+        rectangleRightHorizontalBlock(ModBlocks.MOSSY_RIGHT_LARGE_STONE_BRICK, ModBlocks.MOSSY_LEFT_LARGE_STONE_BRICK);
+
         simpleBlock(ModBlocks.MOSSY_BRICKS.get());
+        stairsBlock(ModBlocks.MOSSY_BRICK_STAIRS.get(), modLoc("block/mossy_bricks"));
+
         simpleBlock(ModBlocks.LARGE_BRICKS.get());
         simpleBlock(ModBlocks.MOSSY_LARGE_BRICKS.get());
+        stairsBlock(ModBlocks.LARGE_BRICK_STAIRS.get(), modLoc("block/large_bricks"));
+        stairsBlock(ModBlocks.MOSSY_LARGE_BRICK_STAIRS.get(), modLoc("block/mossy_large_bricks"));
         simpleBlock(ModBlocks.SQUARE_BRICK.get());
         simpleBlock(ModBlocks.MOSSY_SQUARE_BRICK.get());
         rectangleLeftHorizontalBlock(ModBlocks.LEFT_LARGE_BRICK, ModBlocks.RIGHT_LARGE_BRICK);
         rectangleRightHorizontalBlock(ModBlocks.RIGHT_LARGE_BRICK, ModBlocks.LEFT_LARGE_BRICK);
+        rectangleLeftHorizontalBlock(ModBlocks.MOSSY_LEFT_LARGE_BRICK, ModBlocks.MOSSY_RIGHT_LARGE_BRICK);
+        rectangleRightHorizontalBlock(ModBlocks.MOSSY_RIGHT_LARGE_BRICK, ModBlocks.MOSSY_LEFT_LARGE_BRICK);
 
+        simpleBlock(ModBlocks.COBBLESTONE_BRICK.get());
+        simpleBlock(ModBlocks.MOSSY_COBBLESTONE_BRICK.get());
+        simpleBlock(ModBlocks.GRAVEL_BRICK.get());
     }
 
     private void blockWithItem(RegistryObject<Block> blockRegistryObject) {
@@ -303,8 +360,23 @@ public class ModBlockStateProvider extends BlockStateProvider {
         }, WallRingBlock.WATERLOGGED, WaterloggedNonCubeFacingBlock.WATERLOGGED);
     }
 
-    public void plateBracketBlock(RegistryObject<Block> block) {
-        ModelFile model = models().getExistingFile(modLoc(ModelProvider.BLOCK_FOLDER + "/plate_bracket_block"));
+    public void plateBracketBlock(RegistryObject<Block> block, ResourceLocation texture) {
+        ModelFile model = models().singleTexture(block.getId().getPath(), modLoc(ModelProvider.BLOCK_FOLDER + "/plate_bracket_block"), "0", texture);
+        allDirectionBlock(block.get(), model);
+    }
+
+    public void anglePlateBracketBlock(RegistryObject<Block> block, ResourceLocation texture) {
+        ModelFile model = models().singleTexture(block.getId().getPath(), modLoc(ModelProvider.BLOCK_FOLDER + "/angle_plate_bracket_block"), "0", texture);
+        facingHalfBlock((FacingHalfBlock) block.get(), model);
+    }
+
+    public void cornerPlateBracketBlock(RegistryObject<Block> block, ResourceLocation texture) {
+        ModelFile model = models().singleTexture(block.getId().getPath(), modLoc(ModelProvider.BLOCK_FOLDER + "/corner_plate_bracket_block"), "0", texture);
+        facingHalfBlock((FacingHalfBlock) block.get(), model);
+    }
+
+    public void allDirectionBlock(RegistryObject<Block> block, String name) {
+        ModelFile model = models().getExistingFile(modLoc(ModelProvider.BLOCK_FOLDER + "/" + name));
         allDirectionBlock(block.get(), model);
     }
 
@@ -340,6 +412,11 @@ public class ModBlockStateProvider extends BlockStateProvider {
     public void heavyGrateBlock(RegistryObject<Block> block, ResourceLocation texture) {
         ModelFile model = models().singleTexture(block.getId().getPath(), modLoc("block/template_heavy_grate_block"), "0", texture);
         myDirectionalBlock((Block)block.get(), (ModelFile)model);
+    }
+
+    public void valveWheelBlock(RegistryObject<Block> block, ResourceLocation texture) {
+        ModelFile model = models().singleTexture(block.getId().getPath(), modLoc("block/template_valve_wheel"), "0", texture);
+        allDirectionBlock((Block)block.get(), (ModelFile)model);
     }
 
     @Deprecated
@@ -568,5 +645,30 @@ public class ModBlockStateProvider extends BlockStateProvider {
 
     private ResourceLocation key(Block block) {
         return ForgeRegistries.BLOCKS.getKey(block);
+    }
+
+    /*
+     * this is based off of vanilla stairs, however, this method does not include SHAPE state,
+     * so there are no state definitions generated for 'corner' blocks where two of the same block meet.
+     */
+    public void facingHalfBlock(FacingHalfBlock block, ModelFile model) { //}, ModelFile stairsInner, ModelFile stairsOuter) {
+        getVariantBuilder(block)
+                .forAllStatesExcept(state -> {
+                    Direction facing = state.getValue(FacingHalfBlock.FACING);
+                    Half half = state.getValue(FacingHalfBlock.HALF);
+                    int yRot = 0;
+                    if (facing != Direction.UP && facing != Direction.DOWN) {
+                        // need to spin it around (models always face north by default instead of south)
+                        yRot = (int) facing.getOpposite().toYRot();
+                    }
+                    yRot %= 360;
+                    boolean uvlock = yRot != 0 || half == Half.BOTTOM; // Don't set uvlock for states that have no rotation
+                    return ConfiguredModel.builder()
+                            .modelFile(model)
+                            .rotationX(half == Half.TOP ? 0 : -90)
+                            .rotationY(yRot)
+                            .uvLock(uvlock)
+                            .build();
+                }, StairBlock.WATERLOGGED);
     }
 }